Finished FromRecord derive macro
