+++ rust/template/ddlog_derive/Cargo.toml
@@ -8,6 +8,6 @@ license = "MIT"
 proc-macro = true
 
 [dependencies]
-syn = "1.0.54"
+syn = "1.0.48"
 quote = "1.0.7"
 proc-macro2 = "1.0.24"
